separate code into 2 modules
