+++ igvjs/static/data/public/gstt1_sample.bam.bai
