+++ igvjs/static/js/init.js
