+++ test_IGV.py
@@ -1,6 +1,5 @@
 import unittest
 import struct
-
 from igvjs import app
 
 class TestIGV(unittest.TestCase):
@@ -17,33 +16,25 @@ class TestIGV(unittest.TestCase):
         self.assertIn(b'<title>IGV - Integrative Genomics Viewer</title>', response.data)
 
     def test_get_data_not_auth(self):
-        response = self.app.get('test/example.vcf')
+        response = self.app.get('static/data/public/gstt1_sample.bam')
         self.assertNotEqual(response, None)
         self.assertEqual(response.status_code, 401)
 
     def test_get_data_auth_disabled(self):
         app.config['USES_OAUTH'] = False
-        response = self.app.get('test/example.vcf')
+        response = self.app.get('static/data/public/gstt1_sample.bam')
         self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Genotype', response.data)
-
-    def test_get_data_from_public_dir(self):
-        app.config['PUBLIC_DIR'] = '/static/data'
-        response = self.app.get('test/example.vcf')
-        self.assertEqual(response.status_code, 200)
-        self.assertIn(b'Genotype', response.data)
 
     def test_get_data_from_private_dir(self):
         app.config['PUBLIC_DIR'] = '/static/js'
-        response = self.app.get('test/example.vcf')
+        response = self.app.get('static/data/public/gstt1_sample.bam')
         self.assertEqual(response.status_code, 401)
         self.assertIn(b'Unauthorized', response.data)
-        self.assertNotIn(b'Genotype', response.data)
 
     def test_get_data_range_header(self):
         start = 25
         size = 100
-        response = self.app.get('test/BufferedReaderTest.bin',
+        response = self.app.get('../test/BufferedReaderTest.bin',
             headers={"Range": "bytes={}-{}".format(start, start+size)})
         for i in range(size):
             expected_value = -128 + start + i
